# Optimize 'json_parse_string' using ARM Neon.