--- a/ext/json/ext/generator/depend
+++ b/ext/json/ext/generator/depend
@@ -1,4 +1,4 @@
 generator.o: generator.c $(srcdir)/../fbuffer/fbuffer.h
 generator.o: generator.c $(srcdir)/../vendor/fpconv.c
 generator.o: generator.c $(srcdir)/../vendor/jeaiii-ltoa.h
-generator.o: generator.c $(srcdir)/simd.h
+generator.o: generator.c $(srcdir)/../simd/simd.h