--- a/ext/json/ext/parser/depend
+++ b/ext/json/ext/parser/depend
@@ -1,1 +1,2 @@
 parser.o: parser.c $(srcdir)/../fbuffer/fbuffer.h
+parser.o: parser.c $(srcdir)/../simd/simd.h