… ext/json/ext/generator/simd.h