This PR uses ARM Neon instructions to optimize the `json_parse_string` function in the parser. 

I did refactor the `simd.h` from the generator to `ext/json/ext/simd/simd.h` and reference it from both the generator and the parser.

Similar to the generator, this can be disabled using the `--disable-parser-use-simd` flag.

## Benchmarks

Machine: Macbook Air M1

`master` branch at commit `41d89748fab7343bfd59e55bc171b97d61ab2bb8`

`neon-simd-parser` commit `e884e5025613db8bf3259286c17ecfa9f48aa537`

```
== Parsing activitypub.json (58160 bytes)
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after   965.000 i/100ms
Calculating -------------------------------------
               after      9.768k (± 0.7%) i/s  (102.38 μs/i) -     49.215k in   5.038831s

Comparison:
              before:     8467.9 i/s
               after:     9767.6 i/s - 1.15x  faster


== Parsing twitter.json (567916 bytes)
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after    88.000 i/100ms
Calculating -------------------------------------
               after    871.681 (± 1.3%) i/s    (1.15 ms/i) -      4.400k in   5.048502s

Comparison:
              before:      809.6 i/s
               after:      871.7 i/s - 1.08x  faster


== Parsing citm_catalog.json (1727030 bytes)
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after    41.000 i/100ms
Calculating -------------------------------------
               after    407.105 (± 7.1%) i/s    (2.46 ms/i) -      2.050k in   5.068731s

Comparison:
              before:      416.8 i/s
               after:      407.1 i/s - same-ish: difference falls within error


== Parsing ohai.json (32444 bytes)
ruby 3.4.2 (2025-02-15 revision d2930f8e7a) +YJIT +PRISM [arm64-darwin24]
Warming up --------------------------------------
               after   980.000 i/100ms
Calculating -------------------------------------
               after      9.622k (± 8.2%) i/s  (103.93 μs/i) -     48.020k in   5.041688s

Comparison:
              before:    10094.9 i/s
               after:     9622.1 i/s - same-ish: difference falls within error
```