@samyron 2025-06-17
[SSE2 Support](https://github.com/samyron/json/pull/5). I can merge that to include it in this PR or I can create a separate PR after this is merged if you prefer. The benchmarks look pretty good.