@samyron 2025-06-18
> What I'd like to figure out, is a way to share the search function, because the logic is exactly the same, and the code non-trivial.
> 
> But that's tricky because the implementation in `generator.c` is currently tied to `FBuffer`.

I think I might know a way to do that, though I'm not entirely sure what the API will be. However, the idea is to create some sort of `vector_iterator` structure. Something like the following:

```
typedef struct _vector_iterator {
  // Does this iterator have another vector's worth of data?
  int (*has_chunk)(struct _vector_iterator *);

  // Return a pointer to the start of the first chunk.
  char * (*ptr)(struct _vector_iterator *);

  // Advance by an entire vector. We may not need this if we have an vector_size function.
  void (*advance)(struct _vector_iterator *);

  // Advance the underlying data structure by some number of bytes.
  void (*advance_by)(struct _vector_iterator *, int);
} vector_iterator;
```

Then we can refactor the search code to use the iterator.

```
static inline FORCE_INLINE bool string_scan_vector(vector_iterator *iter)
{
    while (iter->has_chunk(iter)) {
        char *ptr = iter->ptr(iter);

        // The actual NEON code.
        uint64_t mask = neon_step(ptr);

        if (mask) {
            iter->advance_by(iter, trailing_zeros64(mask) >> 2);
            return true;
        }

        iter->advance(iter);
    }
    
    // Will need to handle the remainder somewhere else. Or make the vector_iterator handle scalar, byte-by-byte code.

    return 0;
}
```

In the parser there might be an implementation something like:

```
typedef struct _parser_vector_iterator {
   vector_iterator *iter;
   JSON_ParserState *state;
   int vector_size;
} parser_vector_iterator;

static int has_chunk(vector_iterator *iter) {
   parser_vector_iterator *i = (parser_vector_iterator *)iter;
   return i->state->cursor <= i->state->end;
}

static char *ptr(vector_iterator *iter) {
   parser_vector_iterator *i = (parser_vector_iterator *)iter;
   return i->state->cursor;
}

void advance_by(struct _vector_iterator *iter, int num_bytes) {
   parser_vector_iterator *i = (parser_vector_iterator *)iter;
   i->state->cursor += num_bytes;
}

void advance(struct _vector_iterator *iter) {
   parser_vector_iterator *i = (parser_vector_iterator *)iter;
   i->state->cursor += i->vector_size;
}

// To create this we'd have something like

static void somewhere(JSON_ParserState *state) {
  parser_vector_iterator iter = {
    .ptr = ptr,
    .has_chunk = has_chunk,
    .advance_to = advance_to,
    .advance = advance,
    .state = state,
    .vector_size = sizeof(uint8x16_t) // Assuming Neon  
  }

  // Use this iterator somewhere... 
}
```

As I type this I'm not entirely sure that's the correct level of abstraction, but hopefully it gets the idea across. We can hopefully have the same interface but the implementations within the Parser and Generator can depend on their own local data structures.