@froydnj 2025-06-19
> With the assumption that x86-64 chips may not have SSE2

All x86-64 chips have SSE2; it comes as part of the base ISA.