@samyron 2025-06-24
The SIMD code should now be shared by the parser and generator.