@byroot 2025-06-24
I'll try to find some time to attempt a solution without function pointers to see if it fare any better.