@samyron 2025-06-24
I should note that on `clang 17` and `gcc 14` on macOS and `clang 16`, `clang 17`, `clang 18` and `gcc 13.3` on Ubuntu inline all of the SIMD and related `iterator` code into `json_parse_string`. A quick check of the generator code on Ubuntu using `gcc 13.3` also shows the code is inlined.