@samyron 2025-06-24
Real world parsing benchmarks on x86-64 using the SSE2 code on a `Intel(R) Core(TM) i7-8850H` using `gcc 13.3`.


```
== Parsing activitypub.json (58160 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after   548.000 i/100ms
Calculating -------------------------------------
               after      5.415k (± 4.0%) i/s  (184.67 μs/i) -     27.400k in   5.068215s

Comparison:
              before:     4135.9 i/s
               after:     5415.0 i/s - 1.31x  faster


== Parsing twitter.json (567916 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after    42.000 i/100ms
Calculating -------------------------------------
               after    425.069 (± 5.2%) i/s    (2.35 ms/i) -      2.142k in   5.052769s

Comparison:
              before:      395.0 i/s
               after:      425.1 i/s - same-ish: difference falls within error


== Parsing citm_catalog.json (1727030 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after    18.000 i/100ms
Calculating -------------------------------------
               after    195.139 (± 6.1%) i/s    (5.12 ms/i) -    972.000 in   5.003306s

Comparison:
              before:      183.7 i/s
               after:      195.1 i/s - same-ish: difference falls within error


== Parsing ohai.json (32444 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after   497.000 i/100ms
Calculating -------------------------------------
               after      4.985k (± 3.3%) i/s  (200.59 μs/i) -     25.347k in   5.089864s

Comparison:
              before:     4486.4 i/s
               after:     4985.3 i/s - 1.11x  faster


Run #2

== Parsing activitypub.json (58160 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after   548.000 i/100ms
Calculating -------------------------------------
               after      5.288k (± 4.7%) i/s  (189.11 μs/i) -     26.852k in   5.089488s

Comparison:
              before:     4208.4 i/s
               after:     5287.9 i/s - 1.26x  faster


== Parsing twitter.json (567916 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after    40.000 i/100ms
Calculating -------------------------------------
               after    442.922 (± 3.8%) i/s    (2.26 ms/i) -      2.240k in   5.065280s

Comparison:
              before:      389.7 i/s
               after:      442.9 i/s - 1.14x  faster


== Parsing citm_catalog.json (1727030 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after    17.000 i/100ms
Calculating -------------------------------------
               after    189.731 (± 3.7%) i/s    (5.27 ms/i) -    952.000 in   5.025374s

Comparison:
              before:      191.2 i/s
               after:      189.7 i/s - same-ish: difference falls within error


== Parsing ohai.json (32444 bytes)
ruby 3.4.1 (2024-12-25 revision 48d4efcb85) +PRISM [x86_64-linux]
Warming up --------------------------------------
               after   447.000 i/100ms
Calculating -------------------------------------
               after      4.511k (± 2.4%) i/s  (221.67 μs/i) -     22.797k in   5.056402s

Comparison:
              before:     4082.0 i/s
               after:     4511.3 i/s - same-ish: difference falls within error
```