@samyron 2025-06-25
There are no more function pointers / iterator structure. The code now uses pointers as output variables to modify the necessary state and pass the match `mask` back to the caller.